geometry-based: evaluate fill rate in all boxes and weight by id

+++ OptAlgoBackendPlay/app/models/algorithm/LocalSearch.scala
@@ -47,5 +47,5 @@ trait SolutionHandler[A <: Solution] {
 
   def getNeighborhood(solution: A): Set[A]
 
-  def evaluate(solution: A): Double
+  def evaluate(solution: A): BigDecimal
 }
